requirements.txt
streamlit>=1.30.0
supabase>=2.0.0
google-generativeai>=0.8.0
